Properly record sizes when two strings are aliased to the same memory.
If you do
   a := "some very long .... string"
   b := a
Then `a` and `b` both point to the same memory.  If you do
   size.Of([]string{a, b})
it should only count the content of the very-long string once.  (But
still count the string header two times, as `a` and `b` each have
their own separate header.)

This commit makes that happen.  Sadly, it means that size.go now
depends on the "unsafe" package, which is not implemented in some
environments (such as appengine.)