Account for the full capacity of slices
A slice's capacity may be larger than its length, either because it was
explicitly created this way or due to the dynamic resizing that happens
when new elements are appended.
The memory occupied by the extra capacity is the size of the element
type multiplied by the number of non-assigned elements.